Logout handling
The application has a bunch of singletons (one per object type).
1) There were holding on to their SmartStore and SmartSync instances.
2) Soups and syncs were only setup at init time

Now, we don't hold on to SmartStore or SmartSync instances and we setup soups and syncs on first access to store/sync manager following a logout.

NB: the app as it is would not support multiple users logged in