Bug fix: search screen should only show results for most recently typed query
Behavior before fix:
If you have many contacts and search for abcd (typing the letters quickly), a search is fired for a, then for ab, then for abc, and finally for abcd. Naturally, the search for abcd ends first, followed by the search for abc, the one for ab and the one for a. So you will end up with the results for a (and the "a" in the search box).

Behavior with fix:
Only results for the most recent search should be shown, other results should be discarded without changing the state of the screen.